correct step to open and close usecase and devices, and rename port for more clear meaning what it is used for.

+++ audio/ril_interface.h
@@ -44,7 +44,7 @@ int ril_close(struct ril_handle *ril);
 
 int ril_set_call_volume(struct ril_handle *ril,
                         enum _SoundType sound_type,
-                        float volume);
+                        int volume);
 
 int ril_set_call_audio_path(struct ril_handle *ril,
                             enum _AudioPath path);
